pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "gitoma"
version = "0.1.0"
description = "AI-powered GitHub repository improvement agent — analyzes, plans, commits, and opens PRs autonomously"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "FabGPT", email = "fabgpt.inbox@gmail.com" }]
keywords = ["github", "ai", "agent", "cli", "automation", "bot"]

dependencies = [
    "typer[all]>=0.12",
    "rich>=13.7",
    "GitPython>=3.1",
    "PyGithub>=2.3",
    "openai>=1.30",
    "toml>=0.10",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "radon>=6.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "mcp[cli]>=1.0",
]

[project.scripts]
gitoma = "gitoma.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.9",
    "types-toml>=0.10",
    "types-requests>=2.31",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gitoma*"]

[tool.setuptools.package-data]
# Ship the cockpit UI assets (CSS/JS/HTML) alongside the Python code.
# Without this, `pip install gitoma` would succeed but `gitoma serve` would
# crash at import time trying to read dashboard.css from site-packages.
"gitoma.ui" = ["assets/*"]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
# cli/__init__.py silences urllib3 warnings before the stdlib/third-party
# imports — the order is intentional so E402 must not be enforced here.
"gitoma/cli/__init__.py" = ["E402"]

[tool.mypy]
python_version = "3.12"
strict = true
